Added permutation and binds lemmas. Need FV/notin dom lemma.
Using binds instead of plain In allows destructing uniquely on the
environment permutations. It remains to prove that if a variable is
not in the environment then it will not be free in the process in
which it types.

This binding/permutation thing will also be needed in the subject
reduction theorem it seems. Unless I can eliminate this.